Changed type for parameter in denoiseSaltPepper

+++ src/denoise/denoiseSaltPepper.h
@@ -29,7 +29,7 @@
 
 namespace prl
 {
-void denoiseSaltPepper(const cv::Mat& inputImage, cv::Mat& outputImage, int kernelSize, int times);
+void denoiseSaltPepper(const cv::Mat& inputImage, cv::Mat& outputImage, int kernelSize, size_t times);
 }
 
 #endif //PRLIB_DENOISESALTPEPPER_H
